legion: a few minor fixes

+++ runtime/legion/legion_config.h
@@ -2064,7 +2064,7 @@ typedef unsigned int legion_local_variable_id_t;
 typedef unsigned int legion_replication_id_t;
 typedef unsigned int legion_shard_id_t;
 typedef unsigned int legion_variant_id_t;
-typedef unsigned long legion_distributed_id_t;
+typedef unsigned long long legion_distributed_id_t;
 typedef unsigned long legion_mapping_tag_id_t;
 typedef unsigned long legion_code_descriptor_id_t;
 typedef unsigned long legion_semantic_tag_t;
